% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_classify_ecg.R
\name{dr_classify_ecg}
\alias{dr_classify_ecg}
\title{Classifies ecg data imported with dr_read_ecgxml}
\usage{
dr_classify_ecg(data)
}
\arguments{
\item{data}{of class data.table}
}
\value{
data.table with new columns added for classification of ecg based on diagnosis statements.
}
\description{
Classifies ecg data imported with dr_read_ecgxml
}
\examples{

}
