% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_read_ecgxml.R
\name{dr_read_ecgxml}
\alias{dr_read_ecgxml}
\title{Read ecg xml files}
\usage{
dr_read_ecgxml(filepath)
}
\arguments{
\item{filepath}{path to file inside quotes}
}
\value{
A dataframe with the ecg data
}
\description{
Extracts ecg data from xml files from
Duke University Hospital MUSE version 9.0.4.16760
}
\examples{

}
